chore: gather TexFileStruct function together
