+++ src/intellisense/texDocumentSymbolProvider.ts
@@ -1,8 +1,14 @@
 import * as vscode from 'vscode';
 import { VirtualFileSystem, parseUri } from '../core/remoteFileSystemProvider';
 import { IntellisenseProvider } from '.';
-import { TexFileStruct, TeXElement, TeXElementType, parseTexFileStruct } from './texDocumentParseUtility';
+import { TeXElement, TeXElementType, genTexElements } from './texDocumentParseUtility';
 import { ROOT_NAME } from '../consts';
+
+type TexFileStruct = {
+    texElements: TeXElement[],
+    childrenPaths: string[],
+    bibFilePaths: string[],
+};
 
 function elementsTypeCast(section: TeXElement): vscode.SymbolKind {
     switch (section.type) {
@@ -39,6 +45,47 @@ function elementsToSymbols(sections: TeXElement[]): vscode.DocumentSymbol[] {
         }
     });
     return symbols;
+}
+
+/*
+    * Convert the file into the struct by:
+    * 1. Construct child, named as Uri.path, from TeXElementType.SubFile
+    * 2. Construct bibFile from TeXElementType.BibFile
+    * 
+    * @param filePath: file path of constructed fileSymbolNode
+    * @param fileContent: file content 
+*/
+async function parseTexFileStruct(fileContent:string): Promise<TexFileStruct>{ 
+    const childrenPaths = [];
+    const bibFilePaths = [];
+    const texSymbols = await genTexElements(fileContent);
+
+    // BFS: Traverse the texElements and build fileSymbol
+    const queue: TeXElement[] = [...texSymbols];
+    while (queue.length > 0) {
+        const symbol = queue.shift();
+        switch (symbol?.type) {
+            case TeXElementType.BibFile:
+                bibFilePaths.push(symbol.label);
+                break;
+            case TeXElementType.SubFile:
+                const subFilePath = symbol.label?.endsWith('.tex') ? symbol.label : `${symbol.label}.tex`;
+                childrenPaths.push(subFilePath);
+                break;
+            default:
+                break;
+        }
+        // append children to queue
+        symbol?.children.forEach( child => {
+            queue.push(child);
+        });
+    }
+
+    return {
+        texElements: texSymbols,
+        childrenPaths: childrenPaths,
+        bibFilePaths: bibFilePaths,
+    };
 }
 
 class ProjectStructRecord {
